refactor : 댓글 패키지 구조 수정
