+++ playwithme/src/main/java/com/idea5/playwithme/article/Controller/ArticleController.java
@@ -6,14 +6,13 @@ import com.idea5.playwithme.article.dto.ArticleUpdateForm;
 import com.idea5.playwithme.article.service.ArticleService;
 import com.idea5.playwithme.board.domain.Board;
 import com.idea5.playwithme.board.service.BoardService;
-import com.idea5.playwithme.comment.CommentService;
-import com.idea5.playwithme.comment.domain.CommentCreateForm;
-import com.idea5.playwithme.comment.domain.CommentDto;
+import com.idea5.playwithme.comment.service.CommentService;
+import com.idea5.playwithme.comment.dto.CommentCreateForm;
+import com.idea5.playwithme.comment.dto.CommentDto;
 import com.idea5.playwithme.event.domain.Event;
 import com.idea5.playwithme.member.domain.Member;
 import com.idea5.playwithme.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
