+++ playwithme/src/main/resources/application.yml
@@ -1,9 +1,9 @@
 #Database
 spring:
   datasource:
-    url: jdbc:mariadb://localhost:3306/play_with_me?useUnicode=true&characterEncoding=utf8&serverTimeZone=Asia/Seoul
+    url: jdbc:mariadb://localhost:3306/playwithme?useUnicode=true&characterEncoding=utf8&serverTimeZone=Asia/Seoul
     username: root
-    password: '0114'
+    password:
     driver-class-name: org.mariadb.jdbc.Driver
 
   jpa:
