+++ playwithme/src/test/java/com/idea5/playwithme/comment/CommentRepositoryTest.java
@@ -7,10 +7,10 @@ import com.idea5.playwithme.article.service.ArticleService;
 import com.idea5.playwithme.board.domain.Board;
 import com.idea5.playwithme.board.domain.repository.BoardRepository;
 import com.idea5.playwithme.comment.domain.Comment;
+import com.idea5.playwithme.comment.repository.CommentRepository;
 import com.idea5.playwithme.member.domain.Member;
 import com.idea5.playwithme.member.repository.MemberRepository;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
